requirements.txt
# requirements.txt - Зависимости проекта
aiogram==3.4.1
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-dotenv==1.0.0